requirements.txt
pygame==2.6.0
numpy==1.26.0
sounddevice==0.4.6
opencv-python==4.5.5.64
PyOpenGL==3.1.6
python-dotenv==1.0.0
feedparser
aiohttp